refactor(data): 重构设置数据模型和阅读器配置
- 将AppLanguage从data class改为enum class并迁移到data模块
- 统一所有设置枚举类的fromName方法，支持默认值参数
- 添加AppTextAlign到Compose TextAlign的转换方法
- 修改字体预览卡片支持文本对齐、行高、段落缩进等功能
- 更新常规设置页面的语言选择逻辑
- 将浮点数值设置改为整数类型以提高精度
- 新增SettingsItemWithIntSlider和SettingsItemWithFloatSlider组件
- 移除废弃的RangeExtensions工具类
- 修复图片尺寸百分比格式化字符串
- 更新主题模式和各类显示模式的处理逻辑

+++ app/src/main/java/io/github/lycosmic/lithe/presentation/settings/reader/ReaderSettingsEvent.kt
@@ -14,7 +14,7 @@ sealed class ReaderSettingsEvent {
     /**
      * 字体大小改变
      */
-    data class OnFontSizeChange(val fontSize: Float) : ReaderSettingsEvent()
+    data class OnFontSizeChange(val fontSize: Int) : ReaderSettingsEvent()
 
     /**
      * 字体粗细改变
@@ -29,7 +29,7 @@ sealed class ReaderSettingsEvent {
     /**
      * 字体间距改变
      */
-    data class OnLetterSpacingChange(val letterSpacing: Float) : ReaderSettingsEvent()
+    data class OnLetterSpacingChange(val letterSpacing: Int) : ReaderSettingsEvent()
 
     /**
      * 文本对齐方式改变
@@ -39,17 +39,17 @@ sealed class ReaderSettingsEvent {
     /**
      * 行高改变
      */
-    data class OnLineHeightChange(val lineHeight: Float) : ReaderSettingsEvent()
+    data class OnLineHeightChange(val lineHeight: Int) : ReaderSettingsEvent()
 
     /**
      * 段间距改变
      */
-    data class OnParagraphSpacingChange(val paragraphSpacing: Float) : ReaderSettingsEvent()
+    data class OnParagraphSpacingChange(val paragraphSpacing: Int) : ReaderSettingsEvent()
 
     /**
      * 段首缩进改变
      */
-    data class OnParagraphIndentChange(val paragraphIndent: Float) : ReaderSettingsEvent()
+    data class OnParagraphIndentChange(val paragraphIndent: Int) : ReaderSettingsEvent()
 
     /**
      * 图片可见性改变
@@ -69,7 +69,7 @@ sealed class ReaderSettingsEvent {
     /**
      * 图片圆角半径改变
      */
-    data class OnImageCornerRadiusChange(val radius: Float) : ReaderSettingsEvent()
+    data class OnImageCornerRadiusChange(val radius: Int) : ReaderSettingsEvent()
 
     /**
      * 图片对齐方式改变
